Refactored event components.
TH-1029. Refactored SubEvents to EventHierarchy and created new file for event list components.
