+++ src/common/translation/i18n/en.json
@@ -198,13 +198,15 @@
       "title": "Other times of the event"
     },
     "subEvents": {
-      "buttonHide": "Hide",
-      "buttonReadMore": "Read more; {{date}}",
-      "buttonShow": "Show all",
       "title": "Related to the event"
     },
     "superEvent": {
       "title": "Relates to the event"
+    },
+    "relatedEvents": {
+      "buttonHide": "Hide",
+      "buttonReadMore": "Read more",
+      "buttonShow": "Show all"
     },
     "shareLinks": {
       "title": "Share Event"
