+++ src/common/translation/i18n/fi.json
@@ -197,13 +197,15 @@
       "title": "Tapahtuman muut ajat"
     },
     "subEvents": {
-      "buttonHide": "Piilota",
-      "buttonReadMore": "Lue lisää: {{date}}",
-      "buttonShow": "Näytä kaikki",
       "title": "Tapahtumaan liittyvät"
     },
     "superEvent": {
       "title": "Liittyy tapahtumaan"
+    },
+    "relatedEvents": {
+      "buttonHide": "Piilota",
+      "buttonReadMore": "Lue lisää",
+      "buttonShow": "Näytä kaikki"
     },
     "shareLinks": {
       "title": "Jaa tapahtuma"
