+++ src/common/translation/i18n/sv.json
@@ -197,13 +197,15 @@
       "title": "Andra evenemangstider"
     },
     "subEvents": {
-      "buttonHide": "Dölj",
-      "buttonReadMore": "Läs mer: {{date}}",
-      "buttonShow": "Visa alt",
       "title": "Relaterade till evenemanget"
     },
     "superEvent": {
       "title": "Relaterat till evenemanget"
+    },
+    "relatedEvents": {
+      "buttonHide": "Dölj",
+      "buttonReadMore": "Läs mer",
+      "buttonShow": "Visa alt"
     },
     "shareLinks": {
       "title": "Dela detta"
