+++ src/domain/event/eventInfo/eventList/eventList.module.scss
@@ -3,7 +3,7 @@
   padding-left: 2.25rem;
 }
 
-.otherEventTimes {
+.eventList {
   margin-bottom: 1rem;
 
   .toggleButton {
